pyproject.toml
[project]
name = "demo-geocontext"
version = "0.1.0"
description = "Interactive demo for mborne/geocontext"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "fastapi>=0.117.1",
    "gradio>=5.47.0",
    "langchain-community>=0.3.29",
    "langchain-google-vertexai>=2.1.2",
    "langchain-mcp-adapters>=0.1.10",
    "langchain-ollama>=0.3.8",
    "langchain[anthropic,community]>=0.3.27",
    "langgraph-checkpoint-postgres>=2.0.23",
    "langgraph-checkpoint-redis>=0.1.1",
    "langgraph[anthropic]>=0.6.7",
    "psycopg2>=2.9.10",
    "uvicorn>=0.37.0",
]
